Move properties to properties.py
